Refactor cloud tests to a separate test library and CI job
