+++ tests/main.rs
@@ -66,51 +66,6 @@ mod integ_tests {
             .await;
     }
 
-    #[tokio::test]
-    async fn tls_test() {
-        let cloud_addr = env::var("TEMPORAL_CLOUD_ADDRESS");
-        let cloud_key = env::var("TEMPORAL_CLIENT_KEY");
-
-        let (cloud_addr, cloud_key) = if let (Ok(c), Ok(ck)) = (cloud_addr, cloud_key) {
-            if ck.is_empty() {
-                return; // secret not present in github, could be a fork, just skip
-            }
-            (c, ck)
-        } else {
-            // Skip the test
-            return;
-        };
-
-        let client_cert = env::var("TEMPORAL_CLIENT_CERT")
-            .expect("TEMPORAL_CLIENT_CERT must be set")
-            .replace("\\n", "\n")
-            .into_bytes();
-        let client_private_key = cloud_key.replace("\\n", "\n").into_bytes();
-        let sgo = ClientOptionsBuilder::default()
-            .target_url(Url::from_str(&cloud_addr).unwrap())
-            .client_name("tls_tester")
-            .client_version("clientver")
-            .tls_cfg(TlsConfig {
-                client_tls_config: Some(ClientTlsConfig {
-                    client_cert,
-                    client_private_key,
-                }),
-                ..Default::default()
-            })
-            .build()
-            .unwrap();
-        let con = sgo
-            .connect(
-                env::var("TEMPORAL_CLOUD_NAMESPACE").expect("TEMPORAL_CLOUD_NAMESPACE must be set"),
-                None,
-            )
-            .await
-            .unwrap();
-        con.list_workflow_executions(100, vec![], "".to_string())
-            .await
-            .unwrap();
-    }
-
     pub(crate) async fn mk_nexus_endpoint(starter: &mut CoreWfStarter) -> String {
         let client = starter.get_client().await;
         let endpoint = format!("mycoolendpoint-{}", rand_6_chars());
